Enhance roles and permissions management to setup directly in the UI.
Added a migration process that could be reused in the future

+++ frontend/src/pages/audits/edit/sections/sections.js
@@ -202,7 +202,11 @@ export default {
 
     computed: {
         canCreateComment: function() {
-            return userStore.isAllowed('audits:comments:create') 
+            return userStore.isAllowed('audits:comments:create') && this.canManageAuditComments('create')
+        },
+
+        canEditComments: function() {
+            return this.canManageAuditComments('update') || this.canManageAuditComments('delete')
         },
 
         saveButtonState: function() {
@@ -311,6 +315,25 @@ export default {
                 this.saveSuccessTimer = null
             }
             this.saveSuccess = false
+        },
+
+        isCurrentUserAuditMember: function() {
+            const creatorId = this.auditParent.creator?._id || this.auditParent.creator
+            const collaborators = this.auditParent.collaborators || []
+            return creatorId === userStore.id || collaborators.some(collaborator => (collaborator._id || collaborator) === userStore.id)
+        },
+
+        hasExactPermission: function(scope) {
+            return userStore.permissions === '*' || Boolean(userStore.permissions?.includes(scope))
+        },
+
+        canManageAuditComments: function(action) {
+            return this.isCurrentUserAuditMember() || this.hasExactPermission(`audits:comments:${action}-all`)
+        },
+
+        updateSectionIfEditable: function() {
+            if (this.frontEndAuditState === this.AUDIT_VIEW_STATE.EDIT)
+                this.updateSection()
         },
 
         setupDraftRecovery: function() {
@@ -467,6 +490,9 @@ export default {
         },
 
         createComment: function(fieldName, commentId) {
+            if (!this.canCreateComment)
+                return
+
             let comment = {
                 sectionId: this.sectionId,
                 fieldName: fieldName,
@@ -487,7 +513,7 @@ export default {
                 this.fieldHighlighted = fieldName
                 this.focusComment(comment)
                 this.scrollSidebarCommentIntoView(newComment._id, "end")
-                this.updateSection()
+                this.updateSectionIfEditable()
             })
             .catch((err) => {
                 Notify.create({
@@ -500,6 +526,9 @@ export default {
         },
 
         deleteComment: function(comment) {
+            if (!userStore.isAllowed('audits:comments:delete') || !this.canManageAuditComments('delete'))
+                return
+
             this.editComment = null
             let commentId = comment._id || comment.commentId
             AuditService.deleteComment(this.auditId, commentId)
@@ -507,7 +536,7 @@ export default {
                 if (this.focusedComment === commentId)
                     this.fieldHighlighted = ""
                 document.dispatchEvent(new CustomEvent('comment-deleted', { detail: { id: commentId } }))
-                this.updateSection()
+                this.updateSectionIfEditable()
             })
             .catch((err) => {
                 Notify.create({
@@ -520,6 +549,9 @@ export default {
         },
 
         updateComment: function(comment) {
+            if (!userStore.isAllowed('audits:comments:update') || !this.canManageAuditComments('update'))
+                return
+
             if (comment.textTemp)
                 comment.text = comment.textTemp
             if (comment.replyTemp){
@@ -532,7 +564,7 @@ export default {
                 .then(() => {
                     this.editComment = null
                     this.editReply = null
-                    this.updateSection()
+                    this.updateSectionIfEditable()
                 })
                 .catch((err) => {
                     Notify.create({
